+++ mjml/elements/mj_section.py
@@ -1,8 +1,16 @@
 
+from collections import namedtuple
+from decimal import Decimal
+import re
 
 from ._base import BodyComponent
-from ..helpers import suffixCssClasses, strip_unit
-from ..lib import merge_dicts
+from ..helpers import parse_percentage, strip_unit, suffixCssClasses
+from ..lib import merge_dicts, AttrDict
+
+
+__all__ = ['MjSection']
+
+Position = namedtuple('Position', ('x', 'y'))
 
 
 class MjSection(BodyComponent):
@@ -13,6 +21,9 @@ class MjSection(BodyComponent):
             'background-url'   : 'string',
             'background-repeat': 'enum(repeat,no-repeat)',
             'background-size'  : 'string',
+            'background-position'  : 'string',
+            'background-position-x': 'string',
+            'background-position-y': 'string',
             'border'           : 'string',
             'border-bottom'    : 'string',
             'border-left'      : 'string',
@@ -35,6 +46,7 @@ class MjSection(BodyComponent):
         return {
             'background-repeat': 'repeat',
             'background-size'  : 'auto',
+            'background-position': 'top center',
             'direction'        : 'ltr',
             'padding'          : '20px 0',
             'text-align'       : 'center',
@@ -48,7 +60,14 @@ class MjSection(BodyComponent):
         containerWidth = self.context['containerWidth']
         fullWidth = self.isFullWidth()
         if self.getAttribute('background-url'):
-            background = {'background': self.getAttribute('background-url') }
+            background = {
+                'background': self.getBackground(),
+                # background size, repeat and position has to be seperate since
+                # yahoo does not support shorthand background css property
+                'background-position': self.getBackgroundString(),
+                'background-repeat': self.getAttribute('background-repeat'),
+                'background-size': self.getAttribute('background-size'),
+            }
         else:
             background = {
                 'background': self.getAttribute('background-color'),
@@ -92,6 +111,27 @@ class MjSection(BodyComponent):
             },
         }
 
+    def getBackground(self):
+        if self.hasBackground():
+            bg_url = self.getAttribute('background-url')
+            bg_size = self.getAttribute('background-size')
+            bg_parts = [
+                f"url({bg_url})",
+                self.getBackgroundString(),
+                f'/ {bg_size}',
+                self.getAttribute('background-repeat'),
+            ]
+        else:
+            bg_parts = []
+        return makeBackgroundString([
+            self.getAttribute('background-color'),
+            *bg_parts,
+        ])
+
+    def getBackgroundString(self):
+        bg_pos = self.getBackgroundPosition()
+        return f'{bg_pos.posX} {bg_pos.posY}'
+
     def getChildContext(self):
         box = self.getBoxWidths()['box']
         child_context = merge_dicts(self.context, {'containerWidth': f'{box}px'})
@@ -115,6 +155,36 @@ class MjSection(BodyComponent):
             section,
             self.renderAfter()
         ])
+
+    def getBackgroundPosition(self):
+        pos = self.parseBackgroundPosition()
+        return AttrDict({
+            'posX': self.getAttribute('background-position-x') or pos.x,
+            'posY': self.getAttribute('background-position-y') or pos.y,
+        })
+
+    def parseBackgroundPosition(self):
+        posSplit = self.getAttribute('background-position').split(' ')
+        if len(posSplit) == 1:
+            val, = posSplit
+            # here we must determine if x or y was provided ; other will be center
+            if val in ('top', 'bottom'):
+                return Position(x='center', y=val)
+            else:
+                return Position(x=val, y='center')
+        elif len(posSplit) == 2:
+            # x and y can be put in any order in background-position so we need
+            # to determine that based on values
+            val1, val2 = posSplit
+            val1_is_top_bottom = (val1 in ('top', 'bottom'))
+            val2_is_left_right = (val2 in ('left', 'right'))
+            if val1_is_top_bottom or (val1 == 'center' and val2_is_left_right):
+                return Position(x=val2, y=val1)
+            else:
+                return Position(x=val1, y=val2)
+        else:
+            # more than 2 values is not supported, let's treat as default value
+            return Position(x='center', y='top')
 
     def hasBackground(self):
         return bool(self.get_attr('background-url'))
@@ -201,6 +271,131 @@ class MjSection(BodyComponent):
               </tr>
             <![endif]-->'''
 
+    def renderWithBackground(self, content):
+        fullWidth = self.isFullWidth()
+        containerWidth = self.context['containerWidth']
+
+        bgPos = self.getBackgroundPosition()
+        x_attr_to_percentage = {
+            'left'  : '0%',
+            'center': '50%',
+            'right' : '100%',
+        }
+        if bgPos.posX in x_attr_to_percentage:
+            bgPosX = x_attr_to_percentage[bgPos.posX]
+        elif is_percentage(bgPos.posX):
+            bgPosX = bgPos.posX
+        else:
+            bgPosX = '50%'
+
+        y_attr_to_percentage = {
+            'top'    : '0%',
+            'center' : '50%',
+            'bottom' : '100%',
+        }
+        if bgPos.posY in y_attr_to_percentage:
+            bgPosY = y_attr_to_percentage[bgPos.posY]
+        elif is_percentage(bgPos.posY):
+            bgPosY = bgPos.posY
+        else:
+            bgPosY = '0%'
+
+        # this logic is different when using repeat or no-repeat
+        vOriginX, vPosX = self._calc_origin_and_pos('x', bgPosX, bgPosY)
+        vOriginY, vPosY = self._calc_origin_and_pos('y', bgPosX, bgPosY)
+
+        vSizeAttributes = {}
+        # If background size is either cover or contain, we tell VML to keep
+        # the aspect and fill the entire element.
+        if self.get_attr('background-size') in ('cover', 'contain'):
+            is_cover = (self.get_attr('background-size') == 'cover')
+            vSizeAttributes = {
+                'size'  : '1,1',
+                'aspect': 'atleast' if is_cover else 'atmost',
+            }
+        elif self.get_attr('background-size') != 'auto':
+            bgSplit = self.get_attr('background-size').split(' ')
+            if len(bgSplit) == 1:
+                vSizeAttributes = {
+                    'size'  : self.get_attr('background-size'),
+                    'aspect': 'atmost', # reproduces height auto
+                }
+            else:
+                vSizeAttributes = {
+                  'size': ','.join(bgSplit),
+                }
+
+        is_bg_norepeat = (self.get_attr('background-repeat') == 'no-repeat')
+        vmlType = 'frame' if is_bg_norepeat else 'tile'
+        if self.get_attr('background-size') == 'auto':
+            # if no size provided, keep old behavior because outlook can't use
+            # original image size with "frame"
+            vmlType = 'tile'
+            # also ensure that images are still cropped the same way
+            vOriginX, vPosX = (0.5, 0.5)
+            vOriginY, vPosY = (0, 0)
+
+        vrect_style = {'mso-width-percent': '1000'} if fullWidth else {'width': str(containerWidth)}
+        vrect_attrs = self.html_attrs(**{
+            'style'  : vrect_style,
+            'xmlns:v': 'urn:schemas-microsoft-com:vml',
+            'fill'   : 'true',
+            'stroke' : 'false',
+        })
+        vfill_attrs = self.html_attrs(**{
+            'origin': f'{vOriginX}, {vOriginY}',
+            'position': f'{vPosX}, {vPosY}',
+            'src': self.getAttribute('background-url'),
+            'color': self.getAttribute('background-color'),
+            'type': vmlType,
+            **vSizeAttributes,
+        })
+        return f'''
+          <!--[if mso | IE]>
+            <v:rect {vrect_attrs} />
+            <v:textbox style="mso-fit-shape-to-text:true" inset="0,0,0,0">
+            <v:fill {vfill_attrs} />
+          <![endif]-->
+              {content}
+            <!--[if mso | IE]>
+            </v:textbox>
+          </v:rect>
+        <![endif]-->
+        '''
+
+    def _calc_origin_and_pos(self, coordinate, bgPosX, bgPosY):
+        isX = (coordinate == 'x')
+        bgRepeat = (self.getAttribute('background-repeat') == 'repeat')
+        if isX:
+            pos = bgPosX
+            origin = bgPosX
+        else:
+            pos = bgPosY
+            origin = bgPosY
+
+        if is_percentage(pos):
+            # Should be percentage at this point
+            percentage_value = parse_percentage(pos)
+            decimal = int(percentage_value) / Decimal(100)
+            if bgRepeat:
+                pos = decimal
+                origin = decimal
+            else:
+                pos = (-50 + decimal * 100) / 100
+                origin = (-50 + decimal * 100) / 100
+        elif bgRepeat:
+            # top (y) or center (x)
+            if isX:
+                origin, pos = ('0.5', '0.5')
+            else:
+                origin, pos = ('0', '0')
+        else:
+            if isX:
+                origin, pos = ('0', '0')
+            else:
+                origin, pos = ('-0.5', '-0.5')
+        return (origin, pos)
+
     def renderBefore(self):
         containerWidth = self.context['containerWidth']
         containerWidth_int = strip_unit(containerWidth)
@@ -230,4 +425,12 @@ class MjSection(BodyComponent):
               </tr>
             </table>
             <![endif]-->'''
+
+
+def is_percentage(value):
+    match = re.search(r'^\d+(\.\d+)?%$', value)
+    return (match is not None)
+
+def makeBackgroundString(parts):
+    return ' '.join(filter(None, parts))